+++ Eter/AirMageZephyraCrow.cpp
@@ -17,7 +17,7 @@ std::string AirMageZephyraCrow::getDescription() const
 	return m_description;
 }
 
-bool AirMageZephyraCrow::playMage(Board& board, Color color, int8_t x, int8_t y)
+bool AirMageZephyraCrow::playMage(Board& board, Color color, int16_t x, int16_t y)
 {
 	if (board[{x,y}].empty())
 	{
